@@ -12,52 +12,51 @@
   },
   "scripts": {
     "dev": "astro dev",
-    "build": "astro check && tsc --noEmit && astro build",
-    "lint": "astro check && tsc --noEmit",
+    "build": "astro build",
+    "lint": "astro check",
     "preview": "astro preview",
-    "fix": "prettier --write --plugin-search-dir=. ."
+    "fix": "prettier --write . --plugin=prettier-plugin-astro",
+    "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@astrojs/check": "^0.2.1",
-    "@astrojs/mdx": "^1.1.3",
+    "@astrojs/check": "^0.5.5",
+    "@astrojs/mdx": "^2.1.1",
     "@astrojs/prefetch": "^0.4.1",
-    "@astrojs/rss": "^3.0.0",
-    "@astrojs/sitemap": "^3.0.2",
-    "@astrojs/solid-js": "^3.0.2",
-    "@astrojs/tailwind": "^5.0.2",
-    "clsx": "^2.0.0",
-    "markdown-it": "^13.0.2",
+    "@astrojs/rss": "^4.0.5",
+    "@astrojs/sitemap": "^3.1.1",
+    "@astrojs/solid-js": "^4.0.1",
+    "@astrojs/tailwind": "^5.1.0",
+    "@astrojs/ts-plugin": "^1.5.3",
+    "clsx": "^2.1.0",
+    "markdown-it": "^14.0.0",
     "mdast-util-to-string": "^4.0.0",
     "reading-time": "^1.5.0",
-    "sanitize-html": "^2.11.0",
-    "sharp": "^0.32.6",
-    "solid-js": "^1.8.4"
+    "sanitize-html": "^2.12.1",
+    "sharp": "^0.33.2",
+    "solid-js": "^1.8.15"
   },
   "devDependencies": {
+    "@ianvs/prettier-plugin-sort-imports": "^4.1.1",
     "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@tailwindcss/forms": "^0.5.6",
+    "@tailwindcss/forms": "^0.5.7",
     "@tailwindcss/line-clamp": "^0.4.4",
     "@tailwindcss/typography": "^0.5.10",
-    "@types/eslint": "^8.44.6",
-    "@types/markdown-it": "^13.0.5",
-    "@types/node": "^20.8.9",
-    "@types/sanitize-html": "^2.9.3",
-    "@typescript-eslint/eslint-plugin": "^6.9.0",
-    "@typescript-eslint/parser": "^6.9.0",
-    "astro": "3.4.0",
-    "autoprefixer": "^10.4.16",
+    "@types/eslint": "^8.56.3",
+    "@types/markdown-it": "^13.0.7",
+    "@types/node": "^20.11.20",
+    "@types/sanitize-html": "^2.11.0",
+    "@typescript-eslint/eslint-plugin": "^7.0.2",
+    "@typescript-eslint/parser": "^7.0.2",
+    "astro": "4.4.4",
+    "autoprefixer": "^10.4.17",
     "dayjs": "^1.11.10",
-    "eslint": "^8.52.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-astro": "^0.29.1",
-    "eslint-plugin-prettier": "^5.0.1",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
-    "eslint-plugin-tailwindcss": "^3.13.0",
-    "prettier": "^3.0.3",
-    "prettier-plugin-astro": "^0.12.1",
-    "prettier-plugin-tailwindcss": "^0.5.6",
-    "rehype-autolink-headings": "^7.0.0",
-    "tailwindcss": "^3.3.5",
-    "typescript": "^5.2.2"
+    "eslint": "^8.57.0",
+    "eslint-plugin-astro": "^0.31.4",
+    "prettier": "^3.2.5",
+    "prettier-plugin-astro": "^0.13.0",
+    "prettier-plugin-tailwindcss": "^0.5.11",
+    "rehype-autolink-headings": "^7.1.0",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.3.3"
   }
 }
