@@ -8,23 +8,10 @@ module.exports = {
   trailingComma: "all",
   tabWidth: 2,
   plugins: [
-    require.resolve("prettier-plugin-astro"),
-    require.resolve("prettier-plugin-tailwindcss"),
+    "prettier-plugin-astro",
+    "prettier-plugin-tailwindcss",
+    "@ianvs/prettier-plugin-sort-imports",
   ],
-  overrides: [
-    {
-      files: "*.astro",
-      options: {
-        arrowParens: "always",
-        printWidth: 80,
-        singleQuote: false,
-        jsxSingleQuote: false,
-        semi: true,
-        trailingComma: "all",
-        tabWidth: 2,
-        parser: "astro",
-      },
-    },
-  ],
-  tailwindConfig: "./tailwind.config.cjs",
+
+  tailwindConfig: "./tailwind.config.ts",
 };
