@@ -1,3 +1,3 @@
-/* eslint-disable @typescript-eslint/triple-slash-reference */
+// eslint-disable-next-line @typescript-eslint/triple-slash-reference
 /// <reference path="../.astro/types.d.ts" />
 /// <reference types="astro/client" />
